[PRMT-4567] rename test files
